Queries now work with any Number as result type

+++ java/com.sap.sailing.datamining/src/com/sap/sailing/datamining/impl/SpeedInKnotsExtractor.java
@@ -2,15 +2,15 @@ package com.sap.sailing.datamining.impl;
 
 import com.sap.sailing.domain.base.Moving;
 
-public class SpeedInKnotsExtractor extends AbstractExtractor<Moving, Integer> {
+public class SpeedInKnotsExtractor extends AbstractExtractor<Moving, Double> {
 
     public SpeedInKnotsExtractor() {
         super("speed in knots");
     }
 
     @Override
-    public Integer extract(Moving dataEntry) {
-        return (int) dataEntry.getSpeed().getKnots();
+    public Double extract(Moving dataEntry) {
+        return dataEntry.getSpeed().getKnots();
     }
 
 }
